test(core): split edge case tests
